requirements.txt
discord.py
Requests
python-dotenv